pyproject.toml
[tool.poetry]
name = "rpi-kasa-power"
version = "0.1.0"
description = ""
authors = ["Kenneth <Kenneth@DESKTOP-VHTOUHH>"]

[tool.poetry.dependencies]
python = ">=3.7,<4.0"
python-kasa = "^0.4.0"
tabulate = "^0.8.9"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
